requirements.txt
Pillow==10.0.0
reportlab==4.0.4
requests==2.31.0
selenium==4.1.3
